Added entry in change log and made code improvements
